tmp/zag_repoq-finished/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repoq"
version = "3.0.0"
description = "Repository Quality CLI with JSON-LD ontology, VC certificates, and ZAG integration"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "RepoQ Maintainers"}]
dependencies = ["typer>=0.12","rich>=13.7","jinja2>=3.1","pyyaml>=6.0.2"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Development Status :: 5 - Production/Stable",
  "Environment :: Console",
  "Topic :: Software Development :: Quality Assurance",
  "Topic :: Software Development :: Version Control :: Git"
]

[project.optional-dependencies]
full = [
  "jsonschema>=4.22",
  "rdflib>=7.0",
  "pyshacl>=0.23.0",
  "lizard>=1.17",
  "graphviz>=0.20"
]
sign = [ "cryptography>=42.0" ]
dev = [ "pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.6", "black>=24.3" ]

[project.scripts]
repoq = "repoq.cli:app"

[tool.black]
line-length = 100
target-version = ["py39"]
